Move calculator to flow directory - childcare-costs-for-tax-credits
This calculator is specific to the childcare-costs-for-tax-credits flow and so
should be co-located.

* SmartAnswer::Calculators:ChildcareCostCalculator
  -> SmartAnswer::ChildcareCostsForTaxCreditsFlow::Calculator

* SmartAnswer::Calculators::ChildcareCostCalculatorTest
  -> SmartAnswer::ChildcareCostsForTaxCreditsFlow::CalculatorTest

Since these changes do not affect the behaviour and all the regression tests
pass, I've updated the checksums with the new location of the calculator.
